refactor(Social): change props structure and add storybook story
