+++ ui/src/lib/consts.ts
@@ -1,0 +1,15 @@
+import linkedInIMG from '@public/social-linkedin.svg';
+import githubIMG from '@public/social-github.svg';
+
+export const SOCIAL_LINKS = [
+	{
+		href: 'https://www.linkedin.com/in/jinsuasti/',
+		src: linkedInIMG,
+		alt: 'Linkedin Profile',
+	},
+	{
+		href: 'https://github.com/locke189',
+		src: githubIMG,
+		alt: 'Github Repositories',
+	},
+];
