+++ ui/src/ui/Hero.stories.tsx
@@ -1,25 +1,21 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { Hero } from '@/ui/Hero';
-import { Default as DefaultSocial } from './Social.stories';
 
 export default {
-  component: Hero,
-  title: 'Hero',
-  tags: ['autodocs'],
+	component: Hero,
+	title: 'Hero',
+	tags: ['autodocs'],
 } satisfies Meta<typeof Hero>;
 
 export type Story = StoryObj<typeof Hero>;
 
 export const Default: Story = {
-  args: {
-    title: (
-      <>
-        Hi, I&apos;m <span className="text-green-700">Juan Insuasti</span>!
-      </>
-    ),
-    description: 'A Frontend Developer turning design concepts into seamless digital experiences.',
-    social: {
-      socialLinks: DefaultSocial.args?.socialLinks,
-    },
-  },
+	args: {
+		title: (
+			<>
+				Hi, I&apos;m <span className="text-green-700">Juan Insuasti</span>!
+			</>
+		),
+		description: 'A Frontend Developer turning design concepts into seamless digital experiences.',
+	},
 };
